% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{covmx_exp}
\alias{covmx_exp}
\alias{covmx_gauss}
\title{Spatial Correlation Models}
\usage{
covmx_exp(distmx, psill, erange, nugget)

covmx_gauss(distmx, psill, erange, nugget)
}
\arguments{
\item{distmx}{is the distance matrix for sampled sites.}

\item{psill}{is the partial sill.}

\item{erange}{is the effective range.}

\item{nugget}{is the nugget effect.}
}
\value{
Covariance Matrix
}
\description{
Note that, currently, only one of these models is implemented
in this package: \code{covmx_exp()}.
}
\section{Functions}{
\itemize{
\item \code{covmx_exp}: Exponential Covariance

\item \code{covmx_gauss}: Gaussian Covariance
}}

