% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trial.R
\name{data_trial}
\alias{data_trial}
\title{Conduct a simulation trial}
\usage{
data_trial(
  seed = sample.int(1e+07, size = 1),
  data,
  var,
  n = 50,
  cortype_est = "Exponential",
  ...
)
}
\arguments{
\item{seed}{is a random seed, by default a random integer.}

\item{data}{is the data frame that contains \code{var} and the coordinates}

\item{var}{is the name of the variable to analyze}

\item{n}{is the number of data points sampled.}

\item{cortype_est}{is the correlation function type used to estimate the
covariance parameters. Available options include
\code{"Exponential"} for the exponential correlation, \code{"Gaussian"} for
the Gaussian correlation, and \code{"Spherical"} for the spherical correlation.
The default is \code{"Exponential"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a data frame with \itemize{
\item \code{approach}, the name of th approach (\code{"Degign"}
or \code{"Model"}).
\item \code{seed}, a non-random integer seed.
\item \code{true_mean}, the realized mean from the entire population.
\item \code{true_var}, the realized variance from the entire population.
\item \code{estimate}, the estimated/predicted mean.
\item \code{sd}, the standard error of the estimated/predicted mean.
\item \code{lb}, a lower 95\% confidence bound
\item \code{ub}, an upper 95\% confidence bound.
}
}
\description{
Take a sample from and subsequently analyze spatially correlated data
using \enumerate{
\item GRTS sample, analyzed with GRTS and FPBK
\item SRS sample, analyzed with traditional SRS and FPBK
}
}
\examples{
data("nla_df")
data_trial(seed = sample.int(1e7, size = 1), data = nla_df,
var = "TOTALHG_RESULT", n = 50)
}
