% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Simulate spatially correlated data}
\usage{
sim_pop(
  N = 100,
  gridded = TRUE,
  cortype = "Exponential",
  psill,
  range,
  nugget,
  resptype = "normal",
  mu = 0,
  ...
)
}
\arguments{
\item{N}{is the total number of data points. For \code{gridded = TRUE},
this must be a perfect square.}

\item{gridded}{\code{TRUE} for gridded sites and \code{FALSE} for
points with random locations.}

\item{cortype}{is the correlation function type used to generate the population.
Available options include
\code{"Exponential"} for the exponential correlation, \code{"Gaussian"} for
the Gaussian correlation, and \code{"Spherical"} for the spherical correlation.
The default is \code{"Exponential"}.}

\item{psill}{is the partial sill.}

\item{range}{is the range.}

\item{nugget}{is the nugget.}

\item{resptype}{the response type of the errors. Available options:
\code{"normal"} for Gaussian errors and \code{"lognormal"} for log Gaussian
errors. The default is Gaussian errors.}

\item{mu}{is the mean of response vector and is \code{0} for all population
units by default.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a data frame with \itemize{
\item \code{x}, a column with the spatial x-coordinates.
\item \code{y}, a column with the spatial y-coordinates.
\item \code{response}, a column with the response variable.
}
}
\description{
simulates a spatially correlated response on either a gridded or
random unit square. The resulting data can then be used in
\code{\link{sim_trial}()}.
}
\examples{
sim_pop(N = 100, gridded = TRUE, cortype = "Exponential", psill = 1,
range = 1, nugget = 0.2, resptype = "normal", mu = 0)
}
