% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trial.R
\name{sim_trial}
\alias{sim_trial}
\title{Conduct a simulation trial}
\usage{
sim_trial(
  seed = sample.int(1e+07, size = 1),
  N = 100,
  n = 50,
  gridded = TRUE,
  cortype = "Exponential",
  psill,
  range,
  nugget,
  cortype_est = "Exponential",
  resptype = "normal",
  mu = 0,
  ...
)
}
\arguments{
\item{seed}{is a random seed, by default a random integer.}

\item{N}{is the total number of data points.
For \code{gridded = TRUE}, this must be a perfect square.}

\item{n}{is the number of data points sampled.}

\item{gridded}{\code{TRUE} for gridded sites and \code{FALSE} for
points with random locations.}

\item{cortype}{is the correlation function type used to generate the population.
Available options include
\code{"Exponential"} for the exponential correlation, \code{"Gaussian"} for
the Gaussian correlation, and \code{"Spherical"} for the spherical correlation.
The default is \code{"Exponential"}.}

\item{psill}{is the partial sill.}

\item{range}{is the range.}

\item{nugget}{is the nugget.}

\item{cortype_est}{is the correlation function type used to estimate the
covariance parameters. Available options include
\code{"Exponential"} for the exponential correlation, \code{"Gaussian"} for
the Gaussian correlation, and \code{"Spherical"} for the spherical correlation.
The default is \code{"Exponential"}.}

\item{resptype}{the response type of the errors. Available options:
\code{"normal"} for Gaussian errors and \code{"lognormal"} for log Gaussian
errors. The default is Gaussian errors.}

\item{mu}{is the mean of response vector and is \code{0} for all population
units by default.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a data frame with \itemize{
\item \code{approach}, the name of th approach (\code{"Degign"}
or \code{"Model"}).
\item \code{seed}, a non-random integer seed.
\item \code{true_mean}, the realized mean from the entire population.
\item \code{true_var}, the realized variance from the entire population.
\item \code{estimate}, the estimated/predicted mean.
\item \code{sd}, the standard error of the estimated/predicted mean.
\item \code{lb}, a lower 95\% confidence bound
\item \code{ub}, an upper 95\% confidence bound.
}
}
\description{
Take a sample from and subsequently analyze spatially correlated data
using \enumerate{
\item GRTS sample, analyzed with GRTS and FPBK
\item SRS sample, analyzed with traditional SRS and FPBK
}
}
\examples{
sim_trial(seed = sample.int(1e7, size = 1), N = 100, n = 50,
gridded = TRUE, cortype = "Exponential", psill = 1, range = 1,
nugget = 0.2, cortype_est = "Exponential")
}
