% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trial_both.R
\name{sim_trial_both}
\alias{sim_trial_both}
\title{Conduct a simulation trial}
\usage{
sim_trial_both(
  seed = sample.int(1e+07, size = 1),
  N = 100,
  n = 50,
  cortype = "Exponential",
  psill,
  erange,
  nugget,
  cortype_est = "Exponential",
  resptype = "normal"
)
}
\arguments{
\item{seed}{is a random seed, by default a random integer.}

\item{N}{is the total number of data points.
For \code{gridded = TRUE}, this must be a perfect square.}

\item{n}{is the number of data points sampled.}

\item{cortype}{is the true correlation function, which is
\code{"Exponential"} by default.}

\item{psill}{is the partial sill.}

\item{erange}{is the effective range.}

\item{nugget}{is the nugget.}

\item{cortype_est}{is the (possibly misspecified) correlation
function used to estimate/predict the mean, total, or other quantity.}

\item{resptype}{is the response variable type, either \code{
"normal"} or \code{"lognormal"}.}
}
\value{
a data frame with \itemize{
\item \code{approach}, the name of th approach (\code{"Degign"}
or \code{"Model"}).
\item \code{seed}, a non-random integer seed.
\item \code{true_mean}, the realized mean from the entire population.
\item \code{true_var}, the realized variance from the entire population.
\item \code{estimate}, the estimated/predicted mean.
\item \code{sd}, the standard error of the estimated/predicted mean.
\item \code{lb}, a lower 95\% confidence bound
\item \code{ub}, an upper 95\% confidence bound.
}
}
\description{
Take a sample from and subsequently analyze spatially correlated data
using \enumerate{
\item GRTS sample, analyzed with GRTS and FPBK
\item SRS sample, analyzed with traditional SRS and FPBK
}
}
\examples{
sim_trial_both(seed = sample.int(1e7, size = 1), N = 100, n = 50,
cortype = "Exponential", psill = 1, erange = 1,
nugget = 0.2, cortype_est = "Exponential")
}
