% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Simulate spatially correlated data}
\usage{
sim_pop(
  N = 100,
  gridded = TRUE,
  cortype = "Exponential",
  psill,
  erange,
  nugget,
  resptype = "normal",
  ...
)
}
\arguments{
\item{N}{is the total number of data points. For \code{gridded = TRUE},
this must be a perfect square.}

\item{gridded}{\code{TRUE} for gridded sites and \code{FALSE} for
points with random locations.}

\item{cortype}{is the correlation function, which is
\code{"Exponential"} by default.}

\item{psill}{is the partial sill.}

\item{erange}{is the effective range.}

\item{nugget}{is the nugget.}

\item{resptype}{the response type. The default is Gaussian errors.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a data frame with \itemize{
\item \code{x}, a column with the spatial x-coordinates.
\item \code{y}, a column with the spatial y-coordinates.
\item \code{response}, a column with the response variable.
}
}
\description{
simulates a spatially correlated response on either a gridded or
random unit square. The resulting data can then be used in
\code{\link{sim_trial}()}.
}
\examples{
sim_pop(N = 100, gridded = TRUE, cortype = "Exponential", psill = 1, erange = 1, nugget = 0.2)
}
